apps/api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = ""
authors = [
    {name = "Eduardo",email = "santos934510@gmail.com"}
]
requires-python = ">=3.11,<4.0.0"
dependencies = [
    "flask (>=3.1.2,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandas (>=3.0.0,<4.0.0)",
    "fpdf2 (>=2.8.5,<3.0.0)",
    "pdfplumber (>=0.11.9,<0.12.0)",
    "markitdown[docx] (>=0.1.4,<0.2.0)",
    "celery[redis] (>=5.6.2,<6.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "openpyxl (>=3.1.5,<4.0.0)",
    "flask-limiter (>=4.1.1,<5.0.0)",
    "flask-cors (>=6.0.2,<7.0.0)",
    "flask-socketio (>=5.6.0,<6.0.0)",
    "flask-smorest (>=0.46.2,<0.47.0)",
    "marshmallow (>=4.2.2,<5.0.0)"
]

[tool.poetry]
packages = [{include = "app", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[tool.poetry.scripts]
start-api = "src.app.entrypoints.api:main"
start-worker = "src.app.entrypoints.worker:main"